# Files/swagger-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ShortStay API",
    "version": "1.0.0",
    "description": "\n## ShortStay Backend API\n\nA full-featured short-term property rental platform API.\n\n### Authentication\nMost endpoints require a JWT Bearer token. Include it in the `Authorization` header:\n```\nAuthorization: Bearer <your_token>\n```\n\n### Roles\n- **guest** — Can search properties, make bookings, write reviews, raise complaints\n- **host** — Can list/manage properties, manage bookings, receive payouts\n- **admin** — Full system access: user management, property approval, reports\n- **accountant** — Manages payments, refunds, disputes, payouts\n- **verifier** — Conducts property inspections and submits reports\n\n### Google OAuth Flow\n1. Frontend redirects browser to `GET /api/auth/google`\n2. After Google login, backend redirects to `{FRONTEND_URL}/auth/callback?token=...&role=...&user_id=...`\n3. Frontend stores the token from the URL query parameter\n      ",
    "contact": {
      "name": "ShortStay Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter the JWT token obtained from /api/auth/login"
      }
    },
    "schemas": {
...
      "Booking": {
        "type": "object",
        "properties": {
          "booking_id": {
            "type": "integer",
            "example": 1
          },
          "property_id": {
            "type": "integer",
            "example": 5
          },
          "guest_id": {
            "type": "integer",
            "example": 2
          },
          "checkin_date": {
            "type": "string",
            "format": "date",
            "example": "2026-07-01"
          },
          "checkout_date": {
            "type": "string",
            "format": "date",
            "example": "2026-07-05"
          },
          "total_price": {
            "type": "number",
            "example": 300
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "confirmed",
              "cancelled",
              "rejected",
              "expired"
            ],
            "example": "pending"
          },
          "cancellation_reason": {
            "type": "string",
            "example": "Change of plans"
          }
        }
      },
      "CreateBookingRequest": {
        "type": "object",
        "required": [
          "property_id",
          "checkin_date",
          "checkout_date"
        ],
        "properties": {
          "property_id": {
            "type": "integer",
            "example": 5
          },
          "checkin_date": {
            "type": "string",
            "format": "date",
            "example": "2026-07-01"
          },
          "checkout_date": {
            "type": "string",
            "format": "date",
            "example": "2026-07-05"
          }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/api/bookings": {
      "post": {
        "summary": "Make a booking for a property (guest only)",
        "tags": [
          "Bookings"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookingRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Booking created — status is \"pending\" until host approves",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Booking"
                }
              }
            }
          },
          "400": {
            "description": "Property not available for selected dates"
          },
          "403": {
            "description": "Forbidden — guests only"
          }
        }
      }
    },
    "/api/bookings/my": {
      "get": {
        "summary": "Get all bookings made by the logged-in guest",
        "tags": [
          "Bookings"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of guest's bookings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Booking"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — guests only"
          }
        }
      }
    },
    "/api/bookings/{id}/cancel": {
      "put": {
        "summary": "Cancel a booking (guest only — must be own booking)",
        "tags": [
          "Bookings"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Booking ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Booking cancelled"
          },
          "403": {
            "description": "Forbidden or booking cannot be cancelled"
          },
          "404": {
            "description": "Booking not found"
          }
        }
      }
    },
    "/api/bookings/host": {
      "get": {
        "summary": "Get all bookings for the host's properties",
        "tags": [
          "Bookings"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of bookings for host's properties",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Booking"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — hosts only"
          }
        }
      }
    },
    "/api/bookings/{id}/approve": {
      "put": {
        "summary": "Approve a pending booking (host only)",
        "tags": [
          "Bookings"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Booking ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Booking approved"
          },
          "403": {
            "description": "Forbidden — hosts only"
          },
          "404": {
            "description": "Booking not found"
          }
        }
      }
    },
    "/api/bookings/{id}/reject": {
      "put": {
        "summary": "Reject a pending booking (host only)",
        "tags": [
          "Bookings"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Booking ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Booking rejected"
          },
          "403": {
            "description": "Forbidden — hosts only"
          },
          "404": {
            "description": "Booking not found"
          }
        }
      }
    },
    "/api/bookings/{id}": {
      "get": {
        "summary": "Get a specific booking by ID (guest or host involved in the booking)",
        "tags": [
          "Bookings"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Booking ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Booking details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Booking"
                }
              }
            }
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Booking not found"
          }
        }
      }
    },
...
    "/api/admin/reports/bookings": {
      "get": {
        "summary": "Booking trends report (admin only, ?format=csv supported)",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "format",
            "schema": {
              "type": "string",
              "enum": [
                "json",
                "csv"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Booking report"
          }
        }
      }
    },
...
  }
}